tras la eval

+++ ft_strlcat.c
@@ -17,10 +17,10 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int		src_len;
-	int		dst_len;
-	int		i;
-	int		j;
+	size_t	src_len;
+	size_t	dst_len;
+	size_t	i;
+	size_t	j;
 
 	src_len = ft_strlen(src);
 	dst_len = ft_strlen(dst);
